backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
flask-jwt-extended
python-dotenv
psycopg2
sqlalchemy
pyjwt
marshmallow